+++ package.json
@@ -6,6 +6,7 @@
     "prop-types": "^15.8.1",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
     "styled-components": "^5.3.5"
   },
@@ -64,7 +65,6 @@
     "eslint-plugin-react-hooks": "^4.4.0",
     "eslint-plugin-storybook": "^0.5.11",
     "prettier": "^2.6.2",
-    "react-router-dom": "^6.3.0",
     "webpack": "^5.72.0"
   }
 }
