+++ src/App.jsx
@@ -1,113 +1,15 @@
-import React, { memo, useEffect, useState } from 'react';
-import styled from 'styled-components';
+import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 
-import {
-  useCardNumber,
-  useValidDate,
-  useCardOwnerName,
-  useCVC,
-  useCardPassword,
-  useModal,
-} from './hooks';
-import CardInputs from './components/CardInputs';
-import { Button, Card } from './components/common';
+import AddPage from './pages/AddPage';
+import CardListPage from './pages/CardListPage';
 
-import { ReactComponent as Arrow } from './assets/arrow.svg';
-import isValidCardInputs from './utils/validator';
-
-const StyledPage = styled.form`
-  background: #fff;
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column;
-  padding: 30px;
-  width: 400px;
-  height: 757px;
-`;
-
-const Header = styled.div`
-  align-items: center;
-  display: flex;
-  margin-bottom: 25px;
-`;
-
-const Title = styled.span`
-  font-size: 16px;
-  margin-left: 18px;
-`;
-
-const StyledCard = styled(Card)`
-  align-self: center;
-  margin-bottom: 25px;
-`;
-
-const NextButton = styled(Button)`
-  align-self: end;
-`;
-
-function App() {
-  const {
-    cardNumber,
-    handler: setCardNumber,
-    encryptedCardNumber,
-  } = useCardNumber('');
-  const [cardOwnerName, setCardOwnerName] = useCardOwnerName('');
-  const [validDate, setValidDate] = useValidDate('');
-  const [CVC, setCVC] = useCVC('');
-  const [firstPassword, setFirstPassword] = useCardPassword('');
-  const [secondPassword, setSecondPassword] = useCardPassword('');
-  const [isModalOpen, toggleIsModalOpen] = useModal(false);
-  const [isPossible, setIsPossible] = useState(false);
-
-  useEffect(() => {
-    try {
-      isValidCardInputs(
-        cardNumber,
-        validDate,
-        CVC,
-        firstPassword,
-        secondPassword
-      );
-      setIsPossible(true);
-    } catch (e) {
-      setIsPossible(false);
-    }
-  }, [cardNumber, validDate, CVC, firstPassword, secondPassword]);
-
+export default function App() {
   return (
-    <StyledPage>
-      <Header>
-        <Button size="small" content={<Arrow />} />
-        <Title>카드 추가</Title>
-      </Header>
-      <StyledCard
-        bgColor="#ADD8E6"
-        size="medium"
-        name={cardOwnerName}
-        number={encryptedCardNumber.split('-').join(' ')}
-        validDate={validDate}
-      />
-      <CardInputs
-        cardNumber={encryptedCardNumber}
-        setCardNumber={setCardNumber}
-        validDate={validDate}
-        setValidDate={setValidDate}
-        cardOwnerName={cardOwnerName}
-        setCardOwnerName={setCardOwnerName}
-        CVC={CVC}
-        setCVC={setCVC}
-        isModalOpen={isModalOpen}
-        toggleModal={toggleIsModalOpen}
-        firstPassword={firstPassword}
-        setFirstPassword={setFirstPassword}
-        secondPassword={secondPassword}
-        setSecondPassword={setSecondPassword}
-      />
-      {isPossible && (
-        <NextButton color="#04C09E" content="다음" fontWeight="bold" />
-      )}
-    </StyledPage>
+    <Router>
+      <Routes>
+        <Route path="/" element={<CardListPage />} />
+        <Route path="/add" element={<AddPage />} />
+      </Routes>
+    </Router>
   );
 }
-
-export default memo(App);
